+++ packages/ui/components/input-datepicker/src/LionInputDatepicker.js
@@ -12,7 +12,7 @@ import {
   ArrowMixin,
 } from '@lion/ui/overlays.js';
 import { LocalizeMixin } from '@lion/ui/localize-no-side-effects.js';
-import { ScopedElementsMixin } from '../../core/src/lit-element.js';
+import { ScopedElementsMixin } from '../../core/src/ScopedElementsMixin.js';
 import { localizeNamespaceLoader } from './localizeNamespaceLoader.js';
 
 /**
