+++ packages/ui/components/input-file/src/LionInputFile.js
@@ -2,7 +2,7 @@ import { LionField } from '@lion/ui/form-core.js';
 import { LocalizeMixin } from '@lion/ui/localize.js';
 import { css, html } from 'lit';
 import { ifDefined } from 'lit/directives/if-defined.js';
-import { ScopedElementsMixin } from '../../core/src/lit-element.js';
+import { ScopedElementsMixin } from '../../core/src/ScopedElementsMixin.js';
 import { FileHandle, MAX_FILE_SIZE } from './FileHandle.js';
 import { LionSelectedFileList } from './LionSelectedFileList.js';
 import { localizeNamespaceLoader } from './localizeNamespaceLoader.js';
