+++ packages/ui/components/select-rich/src/LionSelectRich.js
@@ -2,7 +2,7 @@ import { LionListbox } from '@lion/ui/listbox.js';
 import { html } from 'lit';
 import { SlotMixin, browserDetection } from '@lion/ui/core.js';
 import { OverlayMixin, withDropdownConfig } from '@lion/ui/overlays.js';
-import { ScopedElementsMixin } from '../../core/src/lit-element.js';
+import { ScopedElementsMixin } from '../../core/src/ScopedElementsMixin.js';
 import { LionSelectInvoker } from './LionSelectInvoker.js';
 
 /**
